+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added a "Server Secret" to make the API a bit more secure.
 - Added "locking" to the records, to allow for blocked access while updating.
 - Added a "batch mode," which greatly speeds up PUT operations.
+- Added support for the address lookup in the sample config file.
 
 **1.0.0.2034** *(June 11, 2018)*